requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.5
redis==6.4.0
celery==5.5.3
boto3==1.40.30
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
faster-whisper==1.2.0
orjson==3.11.3
requests==2.32.3
email-validator==2.2.0
